requirements.txt
# Echo Voice Support - Python Dependencies

# Core
asyncio>=3.4.3

# Audio Processing
pyaudio>=0.2.14
numpy>=1.26.0
soundfile>=0.12.1
pydub>=0.25.1

# STT - Faster Whisper (Local)
faster-whisper>=0.10.0

# LLM - Ollama
ollama>=0.1.0
aiohttp>=3.9.0

# TTS - Edge-TTS
edge-tts>=6.1.0

# Utilities
structlog>=24.0.0
python-dotenv>=1.0.0
